# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Football Training Board API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clubs/me:
    get:
      operationId: getMyClub
      tags: [clubs]
      summary: Get current user's club
      responses:
        "200":
          description: Club details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        "404":
          description: Club not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMyClub
      tags: [clubs]
      summary: Update current user's club
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClubRequest"
      responses:
        "200":
          description: Updated club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
  /clubs/me/members:
    get:
      operationId: listClubMembers
      tags: [clubs]
      summary: List members of current user's club
      responses:
        "200":
          description: List of club members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClubMember"
    post:
      operationId: inviteClubMember
      tags: [clubs]
      summary: Invite a member to the club
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Member invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubMember"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/me/members/{userId}:
    patch:
      operationId: updateClubMemberRole
      tags: [clubs]
      summary: Update a club member's role
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRoleRequest"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubMember"
    delete:
      operationId: removeClubMember
      tags: [clubs]
      summary: Remove a member from the club
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Member removed
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Club:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        foundedYear:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    UpdateClubRequest:
      type: object
      properties:
        name:
          type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        foundedYear:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
    ClubMember:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - joinedAt
    InviteMemberRequest:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
        role:
          type: string
      required:
        - email
        - firstName
        - lastName
        - password
        - role
    UpdateMemberRoleRequest:
      type: object
      properties:
        role:
          type: string
      required:
        - role
...